Basic

Coordinate Types

TightBindingLattice.FractCoord — Type

FractCoord

Fractional coordinates.

Members

whole ::Vector{Int}: Integer part of fractional coordinates

fraction ::Vector{Float64}: [0,1) part of fractional coordinates

TightBindingLattice.fract2carte — Function

fract2carte(latticevectors, fc)

Arguments

latticevectors::AbstractArray{<:Real, 2}: square matrix whose columns are lattice vectors.

fc::FractCoord: fractional coordinates

fract2carte(unitcell, fractcoord)

Arguments

unitcell::UnitCell

fractcoord::FractCoord

TightBindingLattice.carte2fract — Function

carte2fract(latticevectors, cc; tol=√ϵ)

Arguments

latticevectors::AbstractArray{<:Real, 2}: square matrix whose columns are lattice vectors.

cc::CarteCoord: cartesian coordinates

tol::Real=Base.rtoldefault(Float64): tolerance

carte2fract(unitcell, cartecoord; tol=√ϵ)

Arguments

unitcell::UnitCell

cc::CarteCoord

tol::Real=Base.rtoldefault(Float64)

UnitCell

TightBindingLattice.UnitCell — Type

UnitCell{O}

Parameters

O: type of "site". Any type can be used, but we recommend using String or tuple of String and Int for compatibility with JSON.

Fields

latticevectors ::Array{Float64, 2}: Lattice vectors

reducedreciprocallatticevectors ::Array{Float64, 2}: Reduced reciprocal lattice vectors (transpose of inverse of latticevectors)

reciprocallatticevectors ::Array{Float64, 2}: Reciprocal lattice vectors. 2π * reducedreciprocallatticevectors

sites ::Vector{Tuple{T, FractCoord}}: List of sites within unit cell

siteindices ::Dict{T, Int}: Indices of sites

TightBindingLattice.make_unitcell — Function

UnitCell

Construct a one-dimensional lattice.

Arguments

latticeconstant::Real: Lattice constant

SiteType: List of sites

Optional Arguments

tol=√ϵ: Tolerance

UnitCell

Construct an n-dimensional lattice.

Arguments

latticevectors ::AbstractArray{<:AbstractFloat, 2}: Lattice vectors

SiteType::DataType

Optional Arguments

tol=√ϵ: Epsilon

TightBindingLattice.makeunitcell — Function

UnitCell

Construct a one-dimensional lattice.

Arguments

latticeconstant::Real: Lattice constant

SiteType: List of sites

Optional Arguments

tol=√ϵ: Tolerance

UnitCell

Construct an n-dimensional lattice.

Arguments

latticevectors ::AbstractArray{<:AbstractFloat, 2}: Lattice vectors

SiteType::DataType

Optional Arguments

tol=√ϵ: Epsilon

TightBindingLattice.dimension — Function

dimension(fc::FractCoord)

Dimension of the fractional coordinates

Arguments

fc::FractCoord: Fractional coordinates.

dimension(ortho::OrthoCube)

Return the spatial dimension of the orthocube.

dimension(uc)

Spatial dimension of the unit cell.

dimension(arg::TranslationOperation)

Spatial dimension of the translation operation

TightBindingLattice.numsite — Function

numsite(unitcell)

Number of sites of the unit cell.

Arguments

uc ::UnitCell

TightBindingLattice.sitecount — Function

sitecount(unitcell)

Number of sites of the unit cell.

Arguments

uc ::UnitCell

TightBindingLattice.addsite! — Function

addsite!(unitcell, sitename, sitecoord)

Add an site to the unit cell.

Arguments

uc ::UnitCell{T}

sitename ::{T}

sitecoord ::FractCoord

TightBindingLattice.hassite — Function

hassite(unitcell, name)

Test whether the unit cell contains the site of given name.

Arguments

unitcell::UnitCell{O}

name::O

TightBindingLattice.getsite — Function

getsite(unitcell, name)

Get the site (its site name and its fractional coordinates) with the given name.

Arguments

unitcell::UnitCell{O}

name::O

Return

(sitename, fractcoord)

getsite(unitcell, index)

Arguments

unitcell::UnitCell

index::Integer

Return

(sitename, fractcoord)

TightBindingLattice.getsiteindex — Function

getsiteindex(unitcell, name)

Get index of the given site.

Arguments

unitcell::UnitCell{O}

name::O

TightBindingLattice.getsitecoord — Function

getsitecoord(unitcell, name)

Get the fractional coordinates of the site with the given name.

Arguments

uc ::UnitCell{O}

name ::O

Return

fractcoord

getsitecoord(unitcell::UnitCell, index::Integer)

Arguments

unitcell::UnitCell

index::Integer

Return

FractCoord

TightBindingLattice.getsiteindexcoord — Function

getsiteindexcoord(unitcell, name)

Arguments

unitcell::UnitCell{T}

name::T

Return

(index, fractcoord)

TightBindingLattice.getsitename — Function

getsitename(unitcell, index)

Arguments

uc::UnitCell

index::Integer

Return

sitename

Missing docstring.

Missing docstring for TightBindingLattice.getsitecoord. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TightBindingLattice.carte2fract. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TightBindingLattice.fract2carte. Check Documenter's build log for details.

TightBindingLattice.whichunitcell — Function

whichunitcell(unitcell, name, cartecoord; tol=√ϵ)

Return

Vector{Int}: the integer coordinate of the unitcell that the specified site/cartesian coordinate belongs to.

whichunitcell(unitcell, name, fractcoord; tol=√ϵ)

Return

Vector{Int}: the integer coordinate of the unitcell that the specified site/fractional coordinate belongs to.

Missing docstring.

Missing docstring for TightBindingLattice.momentumgrid. Check Documenter's build log for details.

TightBindingLattice.findsiteindex — Function

findsiteindex(unitcell::UnitCell, fc::FractCoord; tol=√ϵ)

Find the index of the site, and the unitcell coordinate at the specified fractional coordinate.

Arguments

unitcell::UnitCell

fc::FractCoord

Returns

(site_index, unitcell_vector), or (-1, []) if not found.

OrthoCube

Missing docstring.

Missing docstring for TightBindingLattice.OrthoCube. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TightBindingLattice.dimension. Check Documenter's build log for details.

TightBindingLattice.volume — Function

volume(ortho::OrthoCube)

Return the signed volume of the orthocube, defined by the determinant of the shape.

TightBindingLattice.isequiv — Function

isequiv(lhs::OrthoCube, rhs::OrthoCube)

Check whether the two orthocubes are equivalent.

Missing docstring.

Missing docstring for TightBindingLattice.find_generators. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TightBindingLattice.find_generators_2d. Check Documenter's build log for details.

TightBindingLattice.generate_coordinates — Function

generate_coordinates(ortho, generators)

Generate a list of coordinates of the orthocube

Arguments

ortho::OrthoCube

generator_translations::AbstractMatrix{<:Integer}

Lattice

TightBindingLattice.Lattice — Type

Lattice{O}

Represent a lattice.

Arguments

unitcell::UnitCell{O}

orthocube::OrthoCube

bravais_coordinates::Vector{Vector{Int}}

supercell::UnitCell{Tuple{O, Vector{Int}}}

Missing docstring.

Missing docstring for TightBindingLattice.make_lattice. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TightBindingLattice.makelattice. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TightBindingLattice.dimension. Check Documenter's build log for details.