Group

AbstractGroup

TightBindingLattice.AbstractGroup — Type

AbstractGroup

Abstract type for abstract groups. Currently the only subtype is FiniteGroup.

FiniteGroup

TightBindingLattice.FiniteGroup — Type

FiniteGroup

Finite group, with elements {1, 2, 3,..., n}. The identity element is always 1. Can be constructed using FiniteGroup(multiplication_table)

Fields

multiplication_table::Matrix{Int}: multiplication table

period_lengths::Vector{Int}: period length (order) of every element

inverses::Vector{Int}: inverse of every element

conjugacy_classes::Vector{Vector{Int}}: conjugacy classes

Examples

julia> using TightBindingLattice

julia> FiniteGroup([1 2; 2 1])
FiniteGroup([1 2; 2 1], [1, 2], [1, 2], [[1], [2]])

TightBindingLattice.element — Method

element(group, idx)

Return the element of index idx. For FiniteGroup, this is somewhat meaningless since the idxth element is idx. The sole purpose of this function is the bounds checking.

TightBindingLattice.elements — Method

elements(group)

Return the elements of the group.

TightBindingLattice.element_name — Method

element_name(group, idx)

Return the name of element at index idx, which is just the string of idx.

TightBindingLattice.element_names — Method

element_names(group)

Return the names of element.

TightBindingLattice.group_order — Method

group_order(group)

Order of group (i.e. number of elements)

TightBindingLattice.group_order — Method

group_order(group, g)

Order of group element (i.e. period length)

TightBindingLattice.period_length — Method

period_length(group, g)

Order of group element (i.e. period length)

TightBindingLattice.group_multiplication_table — Method

group_multiplication_table(group)

Return multiplcation table of the group.

TightBindingLattice.isabelian — Method

isabelian(group)

Check if the group is abelian.