TightBindingLattice

Lattice.

Group has group structure. Examples are FiniteAbelianGroup, GenericGroup

Multiplication table

order (period length) of every element

Symmetry is the "presentation" of group.

Symmetry on the other hand, is a representation of the group structure in some space. Examples are: TranslationSymmetry, PointSymmetry. Members are

group

generators

conjugacy_classes

character_table

irreps

element names

PointSymmetry has additional info

matrix_representations (i.e. representation in units of lattice vectors)

Schoenflies

Hermann Mauguinn

TightBindingLattice.CarteCoord — Type

CarteCoord

Cartesian coordinates. Vector{Float64}.

TightBindingLattice.FractCoord — Type

FractCoord

Fractional coordinates.

Members

whole ::Vector{Int}: Integer part of fractional coordinates

fraction ::Vector{Float64}: [0,1) part of fractional coordinates

TightBindingLattice.Permutation — Type

    Permutation(perms; max_order=2048)

Create a permutation of integers from 1 to n. perms should be a permutation of 1:n.

Arguments

perms: an integer vector containing a permutation of integers from 1 to n

max_order: maximum order

Note

The convention for the permutation is that map[i] gets mapped to i. In other words, map tells you where each element is from.

TightBindingLattice.UnitCell — Type

UnitCell{O}

Parameters

O: type of "orbital". Any type can be used, but we recommend using String or tuple of String and Int for compatibility with JSON.

Members

latticevectors ::Array{Float64, 2}: Lattice vectors

reducedreciprocallatticevectors ::Array{Float64, 2}: Reduced reciprocal lattice vectors (transpose of inverse of latticevectors)

reciprocallatticevectors ::Array{Float64, 2}: Reciprocal lattice vectors. 2π * reducedreciprocallatticevectors

orbitals ::Vector{Tuple{T, FractCoord}}: List of orbitals within unit cell

orbitalindices ::Dict{T, Int}: Indices of orbitals

TightBindingLattice.addorbital! — Method

addorbital!

Add an orbital to the unit cell.

Arguments

uc ::UnitCell{T}

orbitalname ::{T}

orbitalcoord ::FractCoord

TightBindingLattice.carte2fract — Method

carte2fract

Arguments

latticevectors ::AbstractArray{<:AbstractFloat, 2}: square matrix whose columns are lattice vectors.

cc ::CarteCoord: cartesian coordinates

tol ::Real=sqrt(eps(Float64)): tolerance

TightBindingLattice.carte2fract — Method

carte2fract

Arguments

latticevectors ::Array{Float64, 2}

cc ::CarteCoord

TightBindingLattice.dimension — Method

dimension

Dimension of the fractional coordinates

Arguments

fc ::FractCoord: Fractional coordinates.

TightBindingLattice.dimension — Method

dimension

Spatial dimension of the unit cell.

TightBindingLattice.embed — Method

embed(lattice, point_operation)

Embed the simplest version of point operation. (no local unitary operation)

TightBindingLattice.embed — Method

embed(lattice, translation_operation)

Embed the simplest version of integer translation (no mapping between orbitals etc.)

TightBindingLattice.findorbitalindex — Method

findorbitalindex

Returns (orbitalindex, unitcellvector), or (-1, []) if not found.

TightBindingLattice.fract2carte — Method

fract2carte

Arguments

latticevectors ::AbstractArray{<:AbstractFloat, 2}: square matrix whose columns are lattice vectors.

fc ::FractCoord: fractional coordinates

TightBindingLattice.fract2carte — Method

fract2carte

Arguments

latticevectors ::Array{Float64, 2}

fc ::FractCoord

TightBindingLattice.generate_subgroup — Method

generate_subgroup(group, idx)

subgroup generated by generators. ⟨ {g} ⟩

TightBindingLattice.generate_subgroup — Method

generate_subgroup(group, generators)

subgroup generated by generators. ⟨ S ⟩

TightBindingLattice.getorbital — Method

getorbital

Arguments

uc ::UnitCell{T}

index ::Integer

Return

(orbitalname, fractcoord)

TightBindingLattice.getorbital — Method

getorbital

Get the orbital (its orbital name and its fractional coordinates) with the given name.

Arguments

uc ::UnitCell{O}

name ::O

Return

(orbitalname, fractcoord)

TightBindingLattice.getorbitalcoord — Method

getorbitalcoord

Arguments

uc ::UnitCell

idx ::Integer

Return

fractcoord

TightBindingLattice.getorbitalcoord — Method

getorbitalcoord

Get the fractional coordinates of the orbital with the given name.

Arguments

uc ::UnitCell{O}

name ::O

Return

fractcoord

TightBindingLattice.getorbitalindex — Method

getorbitalindex

Get index of the given orbital.

Arguments

uc ::UnitCell{O}

name ::O

TightBindingLattice.getorbitalindexcoord — Method

getorbitalindexcoord

Arguments

uc ::UnitCell{T}

name ::T

Return

(index, fractcoord)

TightBindingLattice.getorbitalname — Method

getorbitalname

Arguments

uc ::UnitCell

index ::Integer

Return

orbitalname

TightBindingLattice.group_inverse — Method

group_inverse(group, g)

Get inverse of element g

TightBindingLattice.group_inverse — Method

group_inverse(group)

Get a function which gives inverse

TightBindingLattice.group_isomorphism — Method

group_isomorphism(group1, group2)

Find the isomorphism ϕ: G₁ → G₂. Return nothing if not isomorphic.

TightBindingLattice.group_multiplication_table — Method

group_multiplication_table(elements, product=(*))

Generate a multiplication table from elements with product.

TightBindingLattice.group_order — Method

group_order(group, g)

Order of group element (i.e. period length)

TightBindingLattice.group_order — Method

group_order(group)

Order of group (i.e. number of elements)

TightBindingLattice.group_product — Method

group_product(group, lhs, rhs)

TightBindingLattice.hasorbital — Method

hasorbital{T}

Test whether the unit cell contains the orbital of given name.

Arguments

uc ::UnitCell{O}

name ::O

TightBindingLattice.issubgroup — Method

issubgroup(mtab, subset)

TightBindingLattice.linpath — Method

momentumpath

Generate a list of momenta

Arguments

anchorpoints

(Optional) nseg - number of points in each segment

TightBindingLattice.little_group — Method

Generate a little group with given elements. The elements of the little group, which may be sparse, are compressed into consecutive integers.

TightBindingLattice.little_group_elements — Method

psym compatible with hypercube

TightBindingLattice.little_symmetry_iso — Method

little_symmetry_iso(tsym, tsym_irrep_index, psym)

Find little symmetry using group isomorphism

TightBindingLattice.make_unitcell — Method

UnitCell

Construct an n-dimensional lattice.

Arguments

latticevectors ::AbstractArray{<:AbstractFloat, 2}: Lattice vectors

OrbitalType::DataType

Optional Arguments

tol=sqrt(eps(Float64)): Epsilon

TightBindingLattice.make_unitcell — Method

UnitCell

Construct a one-dimensional lattice.

Arguments

latticeconstant ::Float64: Lattice constant

OrbitalType: List of orbitals

Optional Arguments

tol=sqrt(eps(Float64)): Tolerance

TightBindingLattice.minimal_generating_set — Function

minimal_generating_set

TightBindingLattice.momentumgrid — Method

momentumgrid

Generate an n-dimensional grid of momenta of given shape

TightBindingLattice.momentumpath — Method

momentumpath

The anchorpoints are given in units of the reciprocal lattice vectors.

TightBindingLattice.numorbital — Method

numorbital

Number of orbitals of the unit cell.

Arguments

uc ::UnitCell

TightBindingLattice.orthogonalize — Method

    orthogonalize

Find orthogonal generators of the hypercube, and reorder the coordinates accordingly

TightBindingLattice.whichunitcell — Method

whichunitcell

Return

R ::Vector{Int}: which unit cell the specificied orbital/cartesian coordinates belongs to.

TightBindingLattice.whichunitcell — Method

whichunitcell

Return

R ::Vector{Int}: which unit cell the specificied orbital/cartesian coordinates belongs to.

Base.:* — Method

    *(p1 ::Permutation, p2 ::Permutation)

Multiply the two permutation. NOT THIS: (Return [p2.map[x] for x in p1.map].) BUT THIS: Return [p1.map[x] for x in p2.map].

Examples

julia> using TightBindingLattice

julia> Permutation([2,1,3]) * Permutation([1,3,2])
Permutation([2, 3, 1], 3)

julia> Permutation([1,3,2]) * Permutation([2,1,3])
Permutation([3, 1, 2], 3)

Base.:== — Method

==(lhs::ProductOperation, rhs::ProductOperation)

Test for LITERAL equality.

Base.:^ — Method

^(perm ::Permutation, pow ::Integer)

Exponentiate the permutation.

Examples

julia> using TightBindingLattice

julia> Permutation([2,3,4,1])^2
Permutation([3, 4, 1, 2], 2)

TightBindingLattice.conjugacy_class — Method

conjugacy_class(group, i)

Conjugacy class of the element i.