TightBindingLattice

TightBindingLattice

Implements tight binding

TightBindingLattice.CarteCoord — Type.

CarteCoord

Cartesian coordinates. Vector{Float64}.

source

TightBindingLattice.FractCoord — Type.

FractCoord

Fractional coordinates.

Members

whole ::Vector{Int}: Integer part of fractional coordinates

fraction ::Vector{Float64}: [0,1) part of fractional coordinates

source

TightBindingLattice.Permutation — Type.

Permutation(perms ::AbstractVector{Int}; max_order=2048)

Create a permutation of integers from 1 to n. perms should be a permutation of 1:n.

Arguments

perms: an integer vector containing a permutation of integers from 1 to n

max_order: maximum order

source

TightBindingLattice.UnitCell — Type.

UnitCell{O}

Parameters

O: type of "orbital". Any type can be used, but we recommend using String or tuple of String and Int for compatibility with JSON.

Members

latticevectors ::Array{Float64, 2}: Lattice vectors

reducedreciprocallatticevectors ::Array{Float64, 2}: Reduced reciprocal lattice vectors (transpose of inverse of latticevectors)

reciprocallatticevectors ::Array{Float64, 2}: Reciprocal lattice vectors. 2π * reducedreciprocallatticevectors

orbitals ::Vector{Tuple{T, FractCoord}}: List of orbitals within unit cell

orbitalindices ::Dict{T, Int}: Indices of orbitals

source

TightBindingLattice.addorbital! — Method.

addorbital!

Add an orbital to the unit cell.

Arguments

uc ::UnitCell{T}

orbitalname ::{T}

orbitalcoord ::FractCoord

source

TightBindingLattice.carte2fract — Method.

carte2fract

Arguments

latticevectors ::AbstractArray{<:AbstractFloat, 2}: square matrix whose columns are lattice vectors.

cc ::CarteCoord: cartesian coordinates

tol ::Real=sqrt(eps(Float64)): tolerance

source

TightBindingLattice.carte2fract — Method.

carte2fract

Arguments

latticevectors ::Array{Float64, 2}

cc ::CarteCoord

source

TightBindingLattice.dimension — Method.

dimension

Dimension of the fractional coordinates

Arguments

fc ::FractCoord: Fractional coordinates.

source

TightBindingLattice.dimension — Method.

dimension

Spatial dimension of the unit cell.

source

TightBindingLattice.fract2carte — Method.

fract2carte

Arguments

latticevectors ::AbstractArray{<:AbstractFloat, 2}: square matrix whose columns are lattice vectors.

fc ::FractCoord: fractional coordinates

source

TightBindingLattice.fract2carte — Method.

fract2carte

Arguments

latticevectors ::Array{Float64, 2}

fc ::FractCoord

source

TightBindingLattice.getorbital — Method.

getorbital

Arguments

uc ::UnitCell{T}

index ::Integer

Return

(orbitalname, fractcoord)

source

TightBindingLattice.getorbital — Method.

getorbital

Get the orbital (its orbital name and its fractional coordinates) with the given name.

Arguments

uc ::UnitCell{O}

name ::O

Return

(orbitalname, fractcoord)

source

TightBindingLattice.getorbitalcoord — Method.

getorbitalcoord

Arguments

uc ::UnitCell

idx ::Integer

Return

fractcoord

source

TightBindingLattice.getorbitalcoord — Method.

getorbitalcoord

Get the fractional coordinates of the orbital with the given name.

Arguments

uc ::UnitCell{O}

name ::O

Return

fractcoord

source

TightBindingLattice.getorbitalindex — Method.

getorbitalindex

Get index of the given orbital.

Arguments

uc ::UnitCell{O}

name ::O

source

TightBindingLattice.getorbitalindexcoord — Method.

getorbitalindexcoord

Arguments

uc ::UnitCell{T}

name ::T

Return

(index, fractcoord)

source

TightBindingLattice.getorbitalname — Method.

getorbitalname

Arguments

uc ::UnitCell

index ::Integer

Return

orbitalname

source

TightBindingLattice.hasorbital — Method.

hasorbital{T}

Test whether the unit cell contains the orbital of given name.

Arguments

uc ::UnitCell{O}

name ::O

source

TightBindingLattice.is_compatible — Method.

is_compatible

Check whether the fractional momentum ([0, 1)ᴺ) compatible with the identity translation. i.e. k¹ R¹ + k² R² + ... + kᴺ Rᴺ = 0 (mod 1)

Arguments

fractional_momentum ::AbstractVector{Rational} : k

identity_translation ::AbstractVector{<:Integer} : R

source

TightBindingLattice.linpath — Method.

momentumpath

Generate a list of momenta

Arguments

anchorpoints

(Optional) nseg - number of points in each segment

source

TightBindingLattice.make_supercell — Method.

Returns

new unit cell

embedding function

which takes the unitcell displacement in the orignal displacement

which returns a tuple of (supercell lattice displacement, sublattice subscript)

source

TightBindingLattice.make_unitcell — Method.

UnitCell

Construct an n-dimensional lattice.

Arguments

latticevectors ::AbstractArray{<:AbstractFloat, 2}: Lattice vectors

OrbitalType::DataType

Optional Arguments

tol=sqrt(eps(Float64)): Epsilon

source

TightBindingLattice.make_unitcell — Method.

UnitCell

Construct a one-dimensional lattice.

Arguments

latticeconstant ::Float64: Lattice constant

OrbitalType: List of orbitals

Optional Arguments

tol=sqrt(eps(Float64)): Tolerance

source

TightBindingLattice.momentumgrid — Method.

momentumgrid

Generate an n-dimensional grid of momenta of given shape

source

TightBindingLattice.momentumpath — Method.

momentumpath

The anchorpoints are given in units of the reciprocal lattice vectors.

source

TightBindingLattice.numorbital — Method.

numorbital

Number of orbitals of the unit cell.

Arguments

uc ::UnitCell

source

TightBindingLattice.whichunitcell — Method.

whichunitcell

Return

R ::Vector{Int}: which unit cell the specificied orbital/cartesian coordinates belongs to.

source

TightBindingLattice.whichunitcell — Method.

whichunitcell

Return

R ::Vector{Int}: which unit cell the specificied orbital/cartesian coordinates belongs to.

source

Base.:* — Method.

*(p1 ::Permutation, p2 ::Permutation)

Multiply the two permutation. Return [p2.map[x] for x in p1.map].

Examples

julia> using TightBindingLattice

julia> Permutation([1,3,2]) * Permutation([2,1,3])
Permutation([2, 3, 1], 3)

julia> Permutation([2,1,3]) * Permutation([1,3,2])
Permutation([3, 1, 2], 3)

source

Base.:^ — Method.

^(perm ::Permutation, pow ::Integer)

Exponentiate the permutation.

Examples

julia> using TightBindingLattice

julia> Permutation([2,3,4,1])^2
Permutation([3, 4, 1, 2], 2)

source