TightBindingLattice

Lattice.

Group has group structure. Examples are FiniteAbelianGroup, GenericGroup

Multiplication table

order (period length) of every element

Symmetry is the "presentation" of group.

Symmetry on the other hand, is a representation of the group structure in some space. Examples are: TranslationSymmetry, PointSymmetry. Members are

group

generators

conjugacy_classes

character_table

irreps

element names

PointSymmetry has additional info

matrix_representations (i.e. representation in units of lattice vectors)

Schoenflies

Hermann Mauguinn

TightBindingLattice.CarteCoord — Type

CarteCoord

Cartesian coordinates. Vector{Float64}.

TightBindingLattice.FractCoord — Type

FractCoord

Fractional coordinates.

Members

whole ::Vector{Int}: Integer part of fractional coordinates

fraction ::Vector{Float64}: [0,1) part of fractional coordinates

TightBindingLattice.IdentityOperation — Type

IdentityOperation{S<:Real}

TightBindingLattice.Permutation — Type

    Permutation(perms; max_order=2048)

Create a permutation of integers from 1 to n. perms should be a permutation of 1:n.

Arguments

perms: an integer vector containing a permutation of integers from 1 to n

max_order: maximum order

Note

The convention for the permutation is that map[i] gets mapped to i. In other words, map tells you where each element is from.

TightBindingLattice.PointOperation — Type

PointOperation{S<:Real}

TightBindingLattice.PointSymmetry — Type

PointSymmetry

TightBindingLattice.SitePermutation — Type

SitePermutation

TightBindingLattice.SpaceOperation — Type

SpaceOperation{Tp<:Real, Tt<:Real}

S: r ↦ M ⋅ ( r + R )

TightBindingLattice.SymmetryEmbedding — Type

SymmetryEmbedding{SymmetryType<:AbstractSymmetry}

Members

lattice::Lattice

symmetry::SymmetryType

elements::Vector{SitePermutation}

TightBindingLattice.SymmorphicIrrepComponent — Type

SymmorphicIrrepComponent{S1<:SymmetryOrEmbedding, S2<:SymmetryOrEmbedding}

TightBindingLattice.SymmorphicSymmetry — Type

SymmorphicSymmetry{S1, S2, E}

TightBindingLattice.UnitCell — Type

UnitCell{O}

Parameters

O: type of "orbital". Any type can be used, but we recommend using String or tuple of String and Int for compatibility with JSON.

Members

latticevectors ::Array{Float64, 2}: Lattice vectors

reducedreciprocallatticevectors ::Array{Float64, 2}: Reduced reciprocal lattice vectors (transpose of inverse of latticevectors)

reciprocallatticevectors ::Array{Float64, 2}: Reciprocal lattice vectors. 2π * reducedreciprocallatticevectors

orbitals ::Vector{Tuple{T, FractCoord}}: List of orbitals within unit cell

orbitalindices ::Dict{T, Int}: Indices of orbitals

TightBindingLattice.addorbital! — Method

addorbital!

Add an orbital to the unit cell.

Arguments

uc ::UnitCell{T}

orbitalname ::{T}

orbitalcoord ::FractCoord

TightBindingLattice.apply_operation — Method

apply_operation(identity{S}, coordinate::AbstractArray{S}) where {S<:Real}

Do nothing.

TightBindingLattice.apply_operation — Method

apply_operation(symop::PointOperation{S}, coordinate::AbstractArray{S}) where {S}

Apply point operation to the coordinates.

TightBindingLattice.apply_operation — Method

apply_operation(symop::TranslationOperation{S}, coordinate::AbstractArray{S}) where S

TightBindingLattice.apply_operation — Method

apply_operation(op::SpaceOperation{Tp, Tt}, coord::AbstractArray{<:Union{Tp, Tt}}) where {Tp, Tt}

TightBindingLattice.carte2fract — Method

carte2fract

Arguments

latticevectors ::AbstractArray{<:AbstractFloat, 2}: square matrix whose columns are lattice vectors.

cc ::CarteCoord: cartesian coordinates

tol ::Real=Base.rtoldefault(Float64): tolerance

TightBindingLattice.carte2fract — Method

carte2fract

Arguments

latticevectors ::Array{Float64, 2}

cc ::CarteCoord

TightBindingLattice.dimension — Method

dimension

Dimension of the fractional coordinates

Arguments

fc ::FractCoord: Fractional coordinates.

TightBindingLattice.dimension — Method

dimension

Spatial dimension of the unit cell.

TightBindingLattice.element — Method

element(group, idx)

Return the element of index idx. For FiniteGroup, this is somewhat meaningless since the idxth element is idx. The sole purpose of this function is the bounds checking.

TightBindingLattice.element_name — Method

element_name(group, idx)

Return the name of element at index idx, which is just the string of idx.

TightBindingLattice.element_names — Method

element_names(group)

Return the names of element.

TightBindingLattice.elements — Method

elements(group)

Return the elements of the group.

TightBindingLattice.embed — Method

embed(lattice, point_operation)

Embed the simplest version of point operation. (no local unitary operation)

TightBindingLattice.embed — Method

embed(lattice, sop::SpaceOperation{<:Integer, <:Integer})

TightBindingLattice.embed — Method

embed(lattice::Lattice, ssym::SymmorphicSymmetry)

TightBindingLattice.embed — Method

embed(lattice, translation_operation)

Embed the simplest version of integer translation (no mapping between orbitals etc.)

TightBindingLattice.findorbitalindex — Method

findorbitalindex

Returns (orbitalindex, unitcellvector), or (-1, []) if not found.

TightBindingLattice.findorbitalmap — Method

findorbitalmap(unitcell, point_operation)

Find which orbital gets mapped to which orbital with what lattice displacement. Return a vector of tuples (j, Rj) at index i, representing that the orbital i gets mapped to orbital j with lattice displacement Rj.

TightBindingLattice.findorbitalmap — Method

findorbitalmap(unitcell, point_symmetry)

TightBindingLattice.findorbitalmap — Method

findorbitalmap(unitcell, translation_operation)

Find which orbital gets mapped to which orbital with what lattice displacement. Return a vector of tuples (j, Rj) at index i, representing that the orbital i gets mapped to orbital j with lattice displacement Rj.

The overall transformation will be a combiation of (1) the result here, and (2) Bravais lattice transformation. Since all the integer translations will be taken care of by the Bravais lattice transformation, Rj here is all zero vectors.

TightBindingLattice.findorbitalmap — Method

findorbitalmap(unitcell, translation_symmetry)

TightBindingLattice.fract2carte — Method

fract2carte

Arguments

latticevectors ::AbstractArray{<:AbstractFloat, 2}: square matrix whose columns are lattice vectors.

fc ::FractCoord: fractional coordinates

TightBindingLattice.fract2carte — Method

fract2carte

Arguments

latticevectors ::Array{Float64, 2}

fc ::FractCoord

TightBindingLattice.generate_subgroup — Method

generate_subgroup(group, idx)

subgroup generated by generators. ⟨ {g} ⟩

TightBindingLattice.generate_subgroup — Method

generate_subgroup(group, generators)

subgroup generated by generators. ⟨ S ⟩

TightBindingLattice.generator_elements — Method

generator_elements(sym::SymmorphicSymmetry)

TightBindingLattice.generator_indices — Method

generator_indices(sym::SymmorphicSymmetry)

TightBindingLattice.get_irrep_components — Method

get_irrep_components(sym::SymmorphicSymmetryEmbedding)

TightBindingLattice.get_irrep_components — Method

get_irrep_components(sym::SymmorphicSymmetry)

TightBindingLattice.get_irrep_components — Method

get_irrep_components(sym::SymmetryOrEmbedding)

Return a generator which gives IrrepComponent(sym, irrep_index, irrep_component).

TightBindingLattice.get_irrep_iterator — Method

get_irrep_iterator(sic::IrrepComponent)

Return a generator which gives (element, amplitude), where amplitude is the irrep component of element.

TightBindingLattice.get_irrep_iterator — Method

get_irrep_iterator(ssic::SymmorphicIrrepComponent)

TightBindingLattice.getorbital — Method

getorbital

Arguments

uc ::UnitCell{T}

index ::Integer

Return

(orbitalname, fractcoord)

TightBindingLattice.getorbital — Method

getorbital

Get the orbital (its orbital name and its fractional coordinates) with the given name.

Arguments

uc ::UnitCell{O}

name ::O

Return

(orbitalname, fractcoord)

TightBindingLattice.getorbitalcoord — Method

getorbitalcoord

Arguments

uc ::UnitCell

idx ::Integer

Return

fractcoord

TightBindingLattice.getorbitalcoord — Method

getorbitalcoord

Get the fractional coordinates of the orbital with the given name.

Arguments

uc ::UnitCell{O}

name ::O

Return

fractcoord

TightBindingLattice.getorbitalindex — Method

getorbitalindex

Get index of the given orbital.

Arguments

uc ::UnitCell{O}

name ::O

TightBindingLattice.getorbitalindexcoord — Method

getorbitalindexcoord

Arguments

uc ::UnitCell{T}

name ::T

Return

(index, fractcoord)

TightBindingLattice.getorbitalname — Method

getorbitalname

Arguments

uc ::UnitCell

index ::Integer

Return

orbitalname

TightBindingLattice.group_inverse — Method

group_inverse(group, g)

Get inverse of element/elements g.

TightBindingLattice.group_inverse — Method

group_inverse(group)

Get a function which gives inverse.

TightBindingLattice.group_isomorphism — Method

group_isomorphism(group1, group2)

Find the isomorphism ϕ: G₁ → G₂. Return nothing if not isomorphic.

TightBindingLattice.group_multiplication_table — Method

group_multiplication_table(group)

Return multiplcation table of the group.

TightBindingLattice.group_multiplication_table — Method

group_multiplication_table(elements, product=(*))

Generate a multiplication table from elements with product.

TightBindingLattice.group_order — Method

group_order(group, g)

Order of group element (i.e. period length)

TightBindingLattice.group_order — Method

group_order(group)

Order of group (i.e. number of elements)

TightBindingLattice.group_order — Method

group_order(sic::IrrepComponent)

Get order of the symmetry group of sic.

TightBindingLattice.group_product — Method

group_product(group, lhs, rhs)

TightBindingLattice.group_product — Method

group_product(group)

Return a function which computes the group product.

TightBindingLattice.hasorbital — Method

hasorbital{T}

Test whether the unit cell contains the orbital of given name.

Arguments

uc ::UnitCell{O}

name ::O

TightBindingLattice.isabelian — Method

isabelian(group)

Check if the group is abelian.

TightBindingLattice.isbragg — Method

isbragg(k, t)

Check for Bragg condition at momentum k and translation t. Fractional momentum is normalized to 1, i.e. lies within [0, 1)ⁿ

TightBindingLattice.isbragg — Method

isbragg(tsym, tsym_irrep_index, translations)

TightBindingLattice.isbragg — Method

isbragg(tsym, tsym_irrep_index, translation)

TightBindingLattice.iscompatible — Method

iscompatible(tsic, psym)

TightBindingLattice.iscompatible — Method

iscompatible(lattice, point_operation)

TightBindingLattice.iscompatible — Method

iscompatible(lattice, point_symmetry)

TightBindingLattice.iscompatible — Method

iscompatible(lattice, translation_operation)

TightBindingLattice.iscompatible — Method

iscompatible(lattice, translation_symmetry)

Test whether lattice and the symmetry are compatible. For translation symmetry, this means that the hypercubic lattice for both are the same.

TightBindingLattice.iscompatible — Method

iscompatible(orthocube, operator::PointOperation{<:Integer})

TightBindingLattice.iscompatible — Method

iscompatible(orthocube, symmetry::PointSymmetry)

TightBindingLattice.iscompatible — Method

iscompatible(orthocube, operator::TranslationOperation{<:Integer})

TightBindingLattice.iscompatible — Method

iscompatible(orthocube, symmetry::TranslationSymmetry)

TightBindingLattice.iscompatible — Method

iscompatible(psym::PointSymmetry, tsym::TranslationSymmetry)

TightBindingLattice.iscompatible — Method

iscompatible(tsymbed, tsym_irrep_index, psymbed)

Check whether the point symmetry embedding psymbed is compatible with the translation symmetry irrep defined by tsym_irrep_index and symmetry(tsymbed). In other words, the little group elements

TightBindingLattice.iscompatible — Method

iscompatible(tsymbed, psymbed)

Check whether the point symmetry embedding psymbed is compatible with the translation symmetry embedding tsymbed, i.e. whether they have the same "lattice".

TightBindingLattice.iscompatible — Method

iscompatible(tsym, tsym_irrep_index, point_operation)

TightBindingLattice.iscompatible — Method

iscompatible(tsym, tsym_irrep_index, psym)

TightBindingLattice.iscompatible — Method

iscompatible(tsym::TranslationSymmetry, pop::PointOperation{<:Integer})

TightBindingLattice.iscompatible — Method

iscompatible(tsym::TranslationSymmetry, psym::PointSymmetry)

TightBindingLattice.iscompatible — Method

iscompatible(lattice::Lattice, ssym::SymmorphicSymmetry{S1, S2, E}) where {S1, S2, E}

TightBindingLattice.ishomomorphic — Method

ishomomorphic(group, representation; product=(*), equal=(==))

Check whether representation is homomorphic to group under product and equal, order preserved.

TightBindingLattice.issubgroup — Method

issubgroup(group, subset)

Check whether the given subset is a subgroup of group.

TightBindingLattice.linpath — Method

momentumpath

Generate a list of momenta

Arguments

anchorpoints

(Optional) nseg - number of points in each segment

TightBindingLattice.little_group — Method

little_group(tsic, psym)

Return the FiniteGroup object that corresponds to the little group of psym at tsic.

TightBindingLattice.little_group — Method

little_group(tsym, tsym_irrep_index, psym)

TightBindingLattice.little_group — Method

little_group(tsym, psym, elements)

Generate a little group with given elements. The elements of the little group, which may be sparse, are compressed into consecutive integers.

TightBindingLattice.little_group_elements — Method

little_group_elements(tsic, psym)

Get the little group elements (i.e. indices) of psym corresponding to the irrep of translation symmetry specified by tsic. tsic and psym are either

IrrepComponent{TranslationSymmetry} and PointSymmetry, or

IrrepComponent{SymmetryEmbedding{TranslationSymmetry}} and SymmetryEmbedding{PointSymmetry}

TightBindingLattice.little_group_elements — Method

little_group_elements(tsymbed, tsym_irrep_index, psymbed)

TightBindingLattice.little_group_elements — Method

little_group_elements(tsymbed, psymbed)

TightBindingLattice.little_group_elements — Method

little_group_elements(tsym, tsym_irrep_index, psym)

Return elements of psym compatible with the translation symmetry tsym at irrep tsym_irrep_index

TightBindingLattice.little_group_elements — Method

little_group_elements(tsym, psym)

Return elements of psym compatible with the translation symmetry tsym.

TightBindingLattice.little_symmetry — Method

little_symmetry(tsic, psym)

Return the PointSymmetry object that corresponds to the little group of psym at tsic.

TightBindingLattice.little_symmetry — Method

little_symmetry(tsymbed, tsym_irrep_index, psymbed)

TightBindingLattice.little_symmetry — Method

little_symmetry(tsymbed, psymbed)

TightBindingLattice.little_symmetry — Method

little_symmetry(tsym, tsym_irrep_index, psym)

TightBindingLattice.little_symmetry — Method

little_symmetry(tsym, psym)

TightBindingLattice.little_symmetry_iso — Method

little_symmetry_iso(tsym, tsym_irrep_index, psym)

Find little symmetry using group isomorphism

TightBindingLattice.make_unitcell — Method

UnitCell

Construct an n-dimensional lattice.

Arguments

latticevectors ::AbstractArray{<:AbstractFloat, 2}: Lattice vectors

OrbitalType::DataType

Optional Arguments

tol=√ϵ: Epsilon

TightBindingLattice.make_unitcell — Method

UnitCell

Construct a one-dimensional lattice.

Arguments

latticeconstant ::Float64: Lattice constant

OrbitalType: List of orbitals

Optional Arguments

tol=√ϵ: Tolerance

TightBindingLattice.minimal_generating_set — Method

minimal_generating_set(group)

Get minimally generating set of the finite group.

TightBindingLattice.momentumgrid — Method

momentumgrid

Generate an n-dimensional grid of momenta of given shape

TightBindingLattice.numorbital — Method

numorbital

Number of orbitals of the unit cell.

Arguments

uc ::UnitCell

TightBindingLattice.period_length — Method

period_length(group, g)

Order of group element (i.e. period length)

TightBindingLattice.project — Method

project(psym, projection; tol=√ϵ)

TightBindingLattice.read_point_symmetry — Method

read_point_symmetry(data::AbstractDict)

TightBindingLattice.symmetry_product — Method

symmetry_product(psym::PointSymmetry)

TightBindingLattice.symmetry_product — Method

symmetry_product(tsym::TranslationSymmetry)

TightBindingLattice.symmetry_product — Method

symmetry_product(sym::SymmorphicSymmetry{TranslationSymmetry,PointSymmetry,S}) where {S<:SpaceOperation{<:Integer}}

TightBindingLattice.translation_symmetry_embedding — Method

translation_symmetry_embedding(lattice::Lattice)

TightBindingLattice.whichunitcell — Method

whichunitcell

Return

R ::Vector{Int}: which unit cell the specificied orbital/cartesian coordinates belongs to.

TightBindingLattice.whichunitcell — Method

whichunitcell

Return

R ::Vector{Int}: which unit cell the specificied orbital/cartesian coordinates belongs to.

Base.:* — Method

    *(p1 ::Permutation, p2 ::Permutation)

Multiply the two permutation. NOT THIS: (Return [p2.map[x] for x in p1.map].) BUT THIS: Return [p1.map[x] for x in p2.map].

Examples

julia> using TightBindingLattice

julia> Permutation([2,1,3]) * Permutation([1,3,2])
Permutation([2, 3, 1], 3)

julia> Permutation([1,3,2]) * Permutation([2,1,3])
Permutation([3, 1, 2], 3)

Base.:^ — Method

^(perm ::Permutation, pow ::Integer)

Exponentiate the permutation.

Examples

julia> using TightBindingLattice

julia> Permutation([2,3,4,1])^2
Permutation([3, 4, 1, 2], 2)

TightBindingLattice.conjugacy_class — Method

conjugacy_class(group, i)

Conjugacy class of the element i.