Basics

Coordinate Types

TightBindingLattice.CarteCoord — Type

CarteCoord

Cartesian coordinates. Vector{Float64}.

TightBindingLattice.FractCoord — Type

FractCoord

Fractional coordinates.

Fields

whole::Vector{Int}: Integer part of fractional coordinates

fraction::Vector{Float64}: [0,1) part of fractional coordinates

Functions

TightBindingLattice.FractCoord — Method

FractCoord(w, f)

Arguments

w::AbstractVector{<:Integer}: whole part of the coordinate

f::AbstractVector{<:Real}: fractional part. In [0, 1).

TightBindingLattice.FractCoord — Method

FractCoord(coord)

Create a FractCoord from coord, where whole is determined by fld(x, 1) and fraction by mod(x, 1).

Arguments

coord::AbstractVector{<:Real}

TightBindingLattice.FractCoord — Method

FractCoord(ndim)

Construct a zero FractCoord of ndim dimensions.

Arguments

ndim::Integer: spatial dimensions

TightBindingLattice.fract2carte — Method

fract2carte(latticevectors, fc)

Arguments

latticevectors::AbstractArray{<:Real, 2}: square matrix whose columns are lattice vectors.

fc::FractCoord: fractional coordinates

TightBindingLattice.carte2fract — Method

carte2fract(latticevectors, cc; tol=√ϵ)

Arguments

latticevectors::AbstractArray{<:Real, 2}: square matrix whose columns are lattice vectors.

cc::CarteCoord: cartesian coordinates

tol::Real=Base.rtoldefault(Float64): tolerance

UnitCell

TightBindingLattice.UnitCell — Type

UnitCell{O}

Represent a unitcell of a lattice, which contains sites at fixed locations (does not yet implement multiple orbitals per site). It is recommended to use makeunitcell rather than the constructor to make a UnitCell object.

Parameters

O: type of "site". Any type can be used, but we recommend using String or tuple of String and Int for compatibility with JSON.

Fields

latticevectors::Array{Float64, 2}: Lattice vectors

reducedreciprocallatticevectors::Array{Float64, 2}: Reduced reciprocal lattice vectors (transpose of inverse of latticevectors)

reciprocallatticevectors ::Array{Float64, 2}: Reciprocal lattice vectors. 2π * reducedreciprocallatticevectors

sites::Vector{Tuple{T, FractCoord}}: List of sites within unit cell

siteindices::Dict{T, Int}: Indices of sites

TightBindingLattice.makeunitcell — Function

makeunitcell(latticevectors; SiteType=Any, tol=√ϵ)

Construct an n-dimensional lattice.

Arguments

latticevectors: Lattice vectors. Can be a nested array of lattice vectors, or a two-dimensional array whose columns are lattice vectors, or a real number.

Optional Arguments

SiteType::DataType

tol=√ϵ: Epsilon

Functions

TightBindingLattice.dimension — Function

dimension(fc::FractCoord)

Dimension of the fractional coordinates

Arguments

fc::FractCoord: Fractional coordinates.

dimension(ortho::OrthoCube)

Return the spatial dimension of the orthocube.

dimension(uc)

Spatial dimension of the unit cell.

dimension(lattice)

Spatial dimension of the lattice.

Arguments

lattice::Lattice

dimension(arg::IdentityOperation)

Return the spatial dimension of the identity operation.

dimension(arg::TranslationOperation)

Spatial dimension of the translation operation

dimension(arg::PointOperation)

Return spatial dimension of arg.

dimension(arg::SpaceOperation)

Return the spatial dimension of arg.

dimension(sym::TranslationSymmetry)

Spatial dimension of the translation symmetry

dimension(sym::PointSymmetry)

Spatial dimension of the point symmetry

TightBindingLattice.numsite — Function

numsite(unitcell)

Number of sites of the unit cell.

Arguments

uc ::UnitCell

TightBindingLattice.sitecount — Function

sitecount(unitcell)

Number of sites of the unit cell.

Arguments

uc ::UnitCell

TightBindingLattice.addsite! — Function

addsite!(unitcell, sitename, sitecoord)

Add an site to the unit cell.

Arguments

uc ::UnitCell{T}

sitename ::{T}

sitecoord ::FractCoord

TightBindingLattice.hassite — Function

hassite(unitcell, name)

Test whether the unit cell contains the site of given name.

Arguments

unitcell::UnitCell{O}

name::O

TightBindingLattice.getsite — Function

getsite(unitcell, name)

Get the site (its site name and its fractional coordinates) with the given name.

Arguments

unitcell::UnitCell{O}

name::O

Return

(sitename, fractcoord)

getsite(unitcell, index)

Arguments

unitcell::UnitCell

index::Integer

Return

(sitename, fractcoord)

TightBindingLattice.getsiteindex — Function

getsiteindex(unitcell, name)

Get index of the given site.

Arguments

unitcell::UnitCell{O}

name::O

TightBindingLattice.getsitecoord — Function

getsitecoord(unitcell, name)

Get the fractional coordinates of the site with the given name.

Arguments

uc ::UnitCell{O}

name ::O

Return

fractcoord

getsitecoord(unitcell, index)

Arguments

unitcell::UnitCell

index::Integer

Return

FractCoord

TightBindingLattice.getsiteindexcoord — Function

getsiteindexcoord(unitcell, name)

Arguments

unitcell::UnitCell{T}

name::T

Return

(index, fractcoord)

TightBindingLattice.getsitename — Function

getsitename(unitcell, index)

Arguments

uc::UnitCell

index::Integer

Return

sitename

TightBindingLattice.carte2fract — Method

carte2fract(unitcell, cartecoord; tol=√ϵ)

Arguments

unitcell::UnitCell

cc::CarteCoord

tol::Real=Base.rtoldefault(Float64)

TightBindingLattice.fract2carte — Method

fract2carte(unitcell, fractcoord)

Arguments

unitcell::UnitCell

fractcoord::FractCoord

TightBindingLattice.whichunitcell — Function

whichunitcell(unitcell, name, cartecoord; tol=√ϵ)

Return

Vector{Int}: the integer coordinate of the unitcell that the specified site/cartesian coordinate belongs to.

whichunitcell(unitcell, name, fractcoord; tol=√ϵ)

Return

Vector{Int}: the integer coordinate of the unitcell that the specified site/fractional coordinate belongs to.

TightBindingLattice.momentumgrid — Method

momentumgrid(unitcell, shape)

Generate an n-dimensional grid of momenta of given shape.

Returns an n-dimensional array of the following form: math k[i_1, i_2, \ldots ] = G \cdot ( \frac{i_1}{n_1}, \frac{i_2}{n_2}, \ldots )` where G is the reciprocal lattice vector

TightBindingLattice.findsiteindex — Function

findsiteindex(unitcell::UnitCell, fc::FractCoord; tol=√ϵ)

Find the index of the site, and the unitcell coordinate at the specified fractional coordinate.

Arguments

unitcell::UnitCell

fc::FractCoord

Returns

(site_index, unitcell_vector), or (-1, []) if not found.

OrthoCube

TightBindingLattice.OrthoCube — Type

OrthoCube(shape)

Represent a hypercubic (Bravais) lattice.

Fields

shape_matrix: a matrix whose columns are the lattice vectors of the

inverse_shape_matrix: (shape_matrix)

wrap: periodic wrapping function which takes an integer array, and maps it onto a site in the Bravais lattice. Analogous to fldmod.

TightBindingLattice.dimension — Method

dimension(ortho::OrthoCube)

Return the spatial dimension of the orthocube.

TightBindingLattice.volume — Method

volume(ortho::OrthoCube)

Return the signed volume of the orthocube, defined by the determinant of the shape.

TightBindingLattice.isequiv — Method

isequiv(lhs::OrthoCube, rhs::OrthoCube)

Check whether the two orthocubes are equivalent.

TightBindingLattice.find_generators — Method

find_generators(ortho::OrthoCube)

Find generators of an OrthoCube. For one-dimensional lattice, which is isomorphic to Zₙ, the generator is +1. For two-dimensional lattices, this function invokes find_generators_2d. Higher dimensions are currently not supported.

TightBindingLattice.find_generators_2d — Method

find_generators_2d(ortho::OrthoCube)

Find translation generators of an OrthoCube.

TightBindingLattice.generate_coordinates — Method

generate_coordinates(ortho, generators)

Generate a list of coordinates of the orthocube

Arguments

ortho::OrthoCube

generator_translations::AbstractMatrix{<:Integer}

Lattice

TightBindingLattice.Lattice — Type

Lattice{O}

Represent a lattice.

Arguments

unitcell::UnitCell{O}

orthocube::OrthoCube

bravais_coordinates::Vector{Vector{Int}}

supercell::UnitCell{Tuple{O, Vector{Int}}}

TightBindingLattice.makelattice — Function

makelattice(unitcell, shape, [generators])

Create a lattice with periodic boundary condition, using the unitcell, shape, and translation generators.

Arguments

unitcell::UnitCell{O}

shape_matrix::AbstractMatrix{<:Integer}: shape of the Bravais lattice. This can also be

a single integer, which is equivalent to a identity matrix times the number.

Optional Arguments

generator_translations::AbstractMatrix{<:Integer}

TightBindingLattice.dimension — Method

dimension(lattice)

Spatial dimension of the lattice.

Arguments

lattice::Lattice