Group

AbstractGroup

TightBindingLattice.AbstractGroup — Type

AbstractGroup

Abstract type for abstract groups. Currently the only subtype is FiniteGroup.

FiniteGroup

TightBindingLattice.FiniteGroup — Type

FiniteGroup

Finite group, with elements {1, 2, 3,..., n}. The identity element is always 1. Can be constructed using FiniteGroup(multiplication_table)

Fields

multiplication_table::Matrix{Int}: multiplication table

period_lengths::Vector{Int}: period length (order) of every element

inverses::Vector{Int}: inverse of every element

conjugacy_classes::Vector{Vector{Int}}: conjugacy classes

Examples

julia> using TightBindingLattice

julia> FiniteGroup([1 2; 2 1])
FiniteGroup([1 2; 2 1], [1, 2], [1, 2], [[1], [2]])

TightBindingLattice.element — Method

element(group, idx)

Return the element of index idx. For FiniteGroup, this is somewhat meaningless since the idxth element is idx. The sole purpose of this function is the bounds checking.

TightBindingLattice.elements — Method

elements(group)

Return the elements of the group.

TightBindingLattice.element_name — Method

element_name(group, idx)

Return the name of element at index idx, which is just the string of idx.

TightBindingLattice.element_names — Method

element_names(group)

Return the names of element.

TightBindingLattice.group_order — Method

group_order(group)

Order of group (i.e. number of elements)

TightBindingLattice.group_order — Method

group_order(group, g)

Order of group element (i.e. period length)

TightBindingLattice.period_length — Method

period_length(group, g)

Order of group element (i.e. period length)

TightBindingLattice.group_multiplication_table — Method

group_multiplication_table(group)

Return multiplcation table of the group.

TightBindingLattice.isabelian — Method

isabelian(group)

Check if the group is abelian.

TightBindingLattice.group_product — Method

group_product(group)

Return a function which computes the group product.

TightBindingLattice.group_inverse — Method

group_inverse(group)

Get a function which gives inverse.

TightBindingLattice.group_inverse — Method

group_inverse(group, g)

Get inverse of element/elements g.

TightBindingLattice.conjugacy_class — Method

conjugacy_class(group::FiniteGroup, i::Integer)

Conjugacy class of the element i.

TightBindingLattice.generate_subgroup — Method

generate_subgroup(group::FiniteGroup, idx::Integer)

subgroup generated by generators. ⟨ {idx} ⟩

TightBindingLattice.issubgroup — Method

issubgroup(group, subset)

Check whether the given subset is a subgroup of group.

TightBindingLattice.minimal_generating_set — Method

minimal_generating_set(group)

Get minimally generating set of the finite group.

TightBindingLattice.group_isomorphism — Method

group_isomorphism(group1, group2)

Find the isomorphism ϕ: G₁ → G₂. Return nothing if not isomorphic.

TightBindingLattice.group_multiplication_table — Method

group_multiplication_table(elements, product=(*))

Generate a multiplication table from elements with product.

TightBindingLattice.ishomomorphic — Method

ishomomorphic(group, representation; product=(*), equal=(==))

Check whether representation is homomorphic to group under product and equal, order preserved.

Permutation

TightBindingLattice.Permutation — Type

Permutation(perms; max_order=2048)

Create a permutation of integers from 1 to n. perms should be a permutation of 1:n.

Arguments

perms: an integer vector containing a permutation of integers from 1 to n

max_order: maximum order

Note

The convention for the permutation is that map[i] gets mapped to i. In other words, map tells you where each element is from.

Base.:* — Method

*(p1 ::Permutation, p2 ::Permutation)

Multiply the two permutation. NOT THIS: (Return [p2.map[x] for x in p1.map].) BUT THIS: (Return [p1.map[x] for x in p2.map].)

Examples

julia> using TightBindingLattice

julia> Permutation([2,1,3]) * Permutation([1,3,2])
Permutation([2, 3, 1], 3)

julia> Permutation([1,3,2]) * Permutation([2,1,3])
Permutation([3, 1, 2], 3)

TightBindingLattice.generate_group — Method

generate_group(generators...)

Return a FiniteGroup generated by the generators.

Arguments

generators::Permutation...: generating permutations

TightBindingLattice.isidentity — Method

isidentity(perm::Permutation)

Test whether the permutation is an identity.