Symmetry Irrep Components

IrrepComponent

LatticeTools.IrrepComponent — Type

IrrepComponent{SymmetryType}

Represent an irrep component. Consider a symmetry S, and the i'th irrep Γᵢ, which has D dimensions. The d-dimensional component (1 ≤ d ≤ D) is represented by Irrep(S, i, D). This type serves as a subscript for iterating over all components of all irreps.

Parameters

SymmetryType<:SymmetryOrEmbedding

Fields

symmetry::SymmetryType

irrep_index::Int

irrep_component::Int

GroupTools.group_order — Method

group_order(sic::IrrepComponent)

Get order of the symmetry group of sic.

LatticeTools.get_irrep_components — Method

get_irrep_components(sym::SymmetryOrEmbedding)

Return a generator which gives IrrepComponent(sym, irrep_index, irrep_component).

LatticeTools.get_irrep_iterator — Method

get_irrep_iterator(sic::IrrepComponent)

Return a generator which gives (element, amplitude), where amplitude is the irrep component of element.

Little Group

LatticeTools.little_group_elements — Method

little_group_elements(tsic, psym)

Return little group elements (i.e. indices) of psym corresponding to the irrep of translation symmetry specified by tsic. tsic and psym are either

IrrepComponent{FiniteTranslationSymmetry} and PointSymmetry, or

IrrepComponent{SymmetryEmbedding{FiniteTranslationSymmetry}} and SymmetryEmbedding{PointSymmetry}

LatticeTools.little_group — Method

little_group(tsic, psym)

Return the FiniteGroup object that corresponds to the little group of psym at tsic.

LatticeTools.little_group — Method

little_group(tsic, psymbed)

Return the FiniteGroup object that corresponds to the little group of psymbed at tsic.

LatticeTools.little_symmetry — Method

little_symmetry(tsic, psym)

Return the PointSymmetry object that corresponds to the little group of psym at tsic.

LatticeTools.iscompatible — Method

iscompatible(tsic, psym)

Test whether the point symmetry psym is compatible with the irrep component tsic (i.e. momentum) of the translation symmetry.

SymmorphicIrrepComponent

LatticeTools.SymmorphicIrrepComponent — Type

SymmorphicIrrepComponent{S1<:SymmetryOrEmbedding, S2<:SymmetryOrEmbedding}

Irrep component of a symmorphic symmetry (embedding).

Fields

normal::IrrepComponent{S1}

rest::IrrepComponent{S2}

GroupTools.group_order — Method

group_order(arg::SymmorphicIrrepComponent)

Group order of the underlying symmorphic group of arg.

LatticeTools.get_irrep_components — Method

get_irrep_components(sym::SymmorphicSymmetry)

Return an iterator which iterates over the irrep component of the symmorphic symmetry sym.

LatticeTools.get_irrep_components — Method

get_irrep_components(sym::SymmorphicSymmetryEmbedding)

Return an iterator which iterates over the irrep component of the symmorphic symmetry embedding sym.

LatticeTools.get_irrep_iterator — Method

get_irrep_iterator(ssic::SymmorphicIrrepComponent)

Return an iterator which iterates over the elements, together with their corresponding irrep coefficient of the given irrep component.

Return

[(r⋅n, cᵣ⋅cₙ) for all r for all n]