Symmetry Operation

AbstractSpaceSymmetryOperation

LatticeTools.AbstractSpaceSymmetryOperation — Type

AbstractSpaceSymmetryOperation{S<:Real}

Abstract space symmetry operation, i.e. translation, point, and space operation. See also TranslationOperation, PointOperation, SpaceOperation.

LatticeTools.domaintype — Method

domaintype(arg::AbstractSpaceSymmetryOperation{S}) where {S<:Real}

Domain type of arg, i.e. the type of the coordinates.

TranslationOperation

LatticeTools.TranslationOperation — Type

TranslationOperation{S<:Real}

Represents translation symmetry operation

Fields

displacement: displacement vector

Examples

julia> using LatticeTools

julia> TranslationOperation([1, 2])
TranslationOperation{Int64}([1, 2])

julia> TranslationOperation([0.5, 0.0, 0.5])
TranslationOperation{Float64}([0.5, 0.0, 0.5])

Constructors

LatticeTools.TranslationOperation — Method

TranslationOperation{S}(displacement::AbstractVector{<:Real}) where {S}

Construct a translation operation of displacement on coordinate space of type S.

LatticeTools.TranslationOperation — Method

TranslationOperation{S}(displacement::AbstractVector{<:Real}) where {S}

Construct a translation operation of displacement on coordinate space of type S.

TranslationOperation(displacement::AbstractVector{S}) where {S<:Real}

Construct a translation operation of displacement on coordinate space of type S.

Properties

GroupTools.isidentity — Method

isidentity(t::TranslationOperation)

Return true if the translation operation is an identity, i.e. iszero(t.displacement)

LatticeTools.istranslation — Method

istranslation(arg::TranslationOperation)

Always return true, since arg is already a translation operation.

LatticeTools.ispoint — Method

ispoint(arg::TranslationOperation)

Return true if arg is a point operation. The only way this can be true is when arg is an identity operation.

LatticeTools.dimension — Method

dimension(arg::TranslationOperation)

Spatial dimension of the translation operation

Apply

GroupTools.apply_operation — Method

apply_operation(symop::TranslationOperation{S}, coordinate::AbstractArray{S}) where S

Return the translated coordinate.

PointOperation

LatticeTools.PointOperation — Type

PointOperation{S<:Real}

Represents point symmetry operation

Fields

matrix: rotation/mirror/inversion matrix

Constructors

LatticeTools.PointOperation — Method

PointOperation(matrix::AbstractMatrix{S}) where {S<:Real}

Construct a point operation with matrix.

Properties

GroupTools.isidentity — Method

isidentity(arg::PointOperation)

Check whether arg is identity.

LatticeTools.istranslation — Method

istranslation(arg::PointOperation)

Check whether arg is a translation operation, i.e. identity.

LatticeTools.ispoint — Method

ispoint(arg::PointOperation)

Check whether arg is a point operation. Always true.

LatticeTools.dimension — Method

dimension(arg::PointOperation)

Return spatial dimension of arg.

Apply

GroupTools.apply_operation — Method

apply_operation(symop::PointOperation{S}, coordinate::AbstractArray{S}) where {S}

Apply point operation to the coordinates.

SpaceOperation

LatticeTools.SpaceOperation — Type

SpaceOperation{Tp<:Real, Tt<:Real}

Represent a spatial symmetry operation of the following form:

\[S_{[M, \mathbf{R}]}: \mathbf{r} \mapsto M \cdot ( \mathbf{r} + \mathbf{R} )\]

Fields

matrix: Rotation matrix

displacement: Displacement vector

Constructors

LatticeTools.SpaceOperation — Method

SpaceOperation(matrix::AbstractMatrix{Tp}, displacement::AbstractVector{Tt}) where {Tp<:Real, Tt<:Real}

Construct a space operation by matrix and displacement.

LatticeTools.SpaceOperation — Method

SpaceOperation([point], [translation])

Construct a space operation by point and translation.

Arguments

point::PointOperation{Tp}

translation::TranslationOperation{Tt}

Properties

GroupTools.isidentity — Method

isidentity(arg::SpaceOperation)

Check whether arg is identity (i.e. identity point and identity translation)

LatticeTools.istranslation — Method

istranslation(arg::SpaceOperation)

Check whether arg is a translation operation (i.e., point operation is identity)

LatticeTools.ispoint — Method

ispoint(arg::SpaceOperation)

Check whether arg is a point operation (i.e. translation is identity)

LatticeTools.dimension — Method

dimension(arg::SpaceOperation)

Return the spatial dimension of arg.

Apply

GroupTools.apply_operation — Method

apply_operation(op::SpaceOperation{Tp, Tt}, coord::AbstractArray{<:Union{Tp, Tt}}) where {Tp, Tt}