API

TightBindingLattice.CarteCoord — Type

CarteCoord

Cartesian coordinates. Vector{Float64}.

TightBindingLattice.FiniteGroup — Type

FiniteGroup

Finite group, with elements {1, 2, 3,..., n}. The identity element is always 1.

Fields

multiplication_table::Matrix{Int}: multiplication table

period_lengths::Vector{Int}: period length (order) of every element

inverses::Vector{Int}: inverse of every element

conjugacy_classes::Vector{Vector{Int}}: conjugacy classes

TightBindingLattice.FractCoord — Type

FractCoord

Fractional coordinates.

Members

whole ::Vector{Int}: Integer part of fractional coordinates

fraction ::Vector{Float64}: [0,1) part of fractional coordinates

TightBindingLattice.IdentityOperation — Type

IdentityOperation{S<:Real} <: AbstractSpaceSymmetryOperation{S}

Represents identity (space symmetry) operation

Fields

dimension::Int: dimension of the space on which the identity operation acts

TightBindingLattice.Permutation — Type

    Permutation(perms; max_order=2048)

Create a permutation of integers from 1 to n. perms should be a permutation of 1:n.

Arguments

perms: an integer vector containing a permutation of integers from 1 to n

max_order: maximum order

Note

The convention for the permutation is that map[i] gets mapped to i. In other words, map tells you where each element is from.

TightBindingLattice.PointOperation — Type

PointOperation{S<:Real}

Represents point symmetry operation

Fields

matrix: rotation/mirror/inversion matrix

TightBindingLattice.PointSymmetry — Type

PointSymmetry

TightBindingLattice.SitePermutation — Type

SitePermutation

Represents a permutation of sites as a symmetry operation of a lattice.

Fields

permutation::Permutation: permutation

TightBindingLattice.SpaceOperation — Type

SpaceOperation{Tp<:Real, Tt<:Real}

Represents a spatial symmetry operation of the following form:

\[S: \mathbf{r} \mapsto M \cdot ( \mathbf{r} + \mathbf{R} )\]

Fields

matrix: Rotation matrix

displacement: Displacement vector

TightBindingLattice.SymmetryEmbedding — Type

SymmetryEmbedding{SymmetryType<:AbstractSymmetry}

Members

lattice::Lattice

symmetry::SymmetryType

elements::Vector{SitePermutation}

TightBindingLattice.SymmorphicIrrepComponent — Type

SymmorphicIrrepComponent{S1<:SymmetryOrEmbedding, S2<:SymmetryOrEmbedding}

TightBindingLattice.SymmorphicSymmetry — Type

SymmorphicSymmetry{S1, S2, E}

TightBindingLattice.TranslationOperation — Type

TranslationOperation{S<:Real}

Represents translation symmetry operation

Fields

displacement: displacement vector

Examples

julia> using TightBindingLattice

julia> TranslationOperation([1, 2])
TranslationOperation{Int64}([1, 2])

julia> TranslationOperation([0.5, 0.0, 0.5])
TranslationOperation{Float64}([0.5, 0.0, 0.5])

TightBindingLattice.TranslationSymmetry — Type

TranslationSymmetry

TightBindingLattice.UnitCell — Type

UnitCell{O}

Parameters

O: type of "site". Any type can be used, but we recommend using String or tuple of String and Int for compatibility with JSON.

Fields

latticevectors ::Array{Float64, 2}: Lattice vectors

reducedreciprocallatticevectors ::Array{Float64, 2}: Reduced reciprocal lattice vectors (transpose of inverse of latticevectors)

reciprocallatticevectors ::Array{Float64, 2}: Reciprocal lattice vectors. 2π * reducedreciprocallatticevectors

sites ::Vector{Tuple{T, FractCoord}}: List of sites within unit cell

siteindices ::Dict{T, Int}: Indices of sites

Base.:* — Method

    *(p1 ::Permutation, p2 ::Permutation)

Multiply the two permutation. NOT THIS: (Return [p2.map[x] for x in p1.map].) BUT THIS: Return [p1.map[x] for x in p2.map].

Examples

julia> using TightBindingLattice

julia> Permutation([2,1,3]) * Permutation([1,3,2])
Permutation([2, 3, 1], 3)

julia> Permutation([1,3,2]) * Permutation([2,1,3])
Permutation([3, 1, 2], 3)

Base.:^ — Method

^(perm ::Permutation, pow ::Integer)

Exponentiate the permutation.

Examples

julia> using TightBindingLattice

julia> Permutation([2,3,4,1])^2
Permutation([3, 4, 1, 2], 2)

TightBindingLattice.addsite! — Method

addsite!

Add an site to the unit cell.

Arguments

uc ::UnitCell{T}

sitename ::{T}

sitecoord ::FractCoord

TightBindingLattice.apply_operation — Method

apply_operation(identity{S}, coordinate::AbstractArray{S}) where {S<:Real}

Do nothing.

TightBindingLattice.apply_operation — Method

apply_operation(symop::PointOperation{S}, coordinate::AbstractArray{S}) where {S}

Apply point operation to the coordinates.

TightBindingLattice.apply_operation — Method

apply_operation(symop::TranslationOperation{S}, coordinate::AbstractArray{S}) where S

Return the translated coordinate.

TightBindingLattice.apply_operation — Method

apply_operation(op::SpaceOperation{Tp, Tt}, coord::AbstractArray{<:Union{Tp, Tt}}) where {Tp, Tt}

TightBindingLattice.carte2fract — Method

carte2fract

Arguments

latticevectors ::AbstractArray{<:Real, 2}: square matrix whose columns are lattice vectors.

cc ::CarteCoord: cartesian coordinates

tol ::Real=Base.rtoldefault(Float64): tolerance

TightBindingLattice.carte2fract — Method

carte2fract

Arguments

latticevectors ::Array{Float64, 2}

cc ::CarteCoord

TightBindingLattice.conjugacy_class — Method

conjugacy_class(group::FiniteGroup, i::Integer)

Conjugacy class of the element i.

TightBindingLattice.dimension — Method

dimension

Dimension of the fractional coordinates

Arguments

fc ::FractCoord: Fractional coordinates.

TightBindingLattice.dimension — Method

dimension(arg::TranslationOperation)

Spatial dimension of the translation operation

TightBindingLattice.dimension — Method

dimension

Spatial dimension of the unit cell.

TightBindingLattice.element — Method

element(group, idx)

Return the element of index idx. For FiniteGroup, this is somewhat meaningless since the idxth element is idx. The sole purpose of this function is the bounds checking.

TightBindingLattice.element_name — Method

element_name(group, idx)

Return the name of element at index idx, which is just the string of idx.

TightBindingLattice.element_names — Method

element_names(group)

Return the names of element.

TightBindingLattice.elements — Method

elements(group)

Return the elements of the group.

TightBindingLattice.embed — Method

embed(lattice, point_operation)

Embed the simplest version of point operation. (no local unitary operation)

TightBindingLattice.embed — Method

embed(lattice, sop::SpaceOperation{<:Integer, <:Integer})

TightBindingLattice.embed — Method

embed(lattice::Lattice, ssym::SymmorphicSymmetry)

TightBindingLattice.embed — Method

embed(lattice, translation_operation)

Embed the simplest version of integer translation (no mapping between sites etc.)

TightBindingLattice.findsiteindex — Method

findsiteindex(unitcell::UnitCell, fc::FractCoord; tol=√ϵ)

Find the index of the site, and the unitcell coordinate at the specified fractional coordinate.

Arguments

unitcell::UnitCell

fc::FractCoord

Returns

(site_index, unitcell_vector), or (-1, []) if not found.

TightBindingLattice.findsitemap — Method

findsitemap(unitcell, point_operation)

Find which site gets mapped to which site with what lattice displacement. Return a vector of tuples (j, Rj) at index i, representing that the site i gets mapped to site j with lattice displacement Rj. If the unit cell is not compatible with the point operation, return nothing.

TightBindingLattice.findsitemap — Method

findsitemap(unitcell, point_symmetry)

Find the sitemap for every element of the point symmetry. Return nothing if the unit cell is incompatible with any element of the point symmetry.

TightBindingLattice.findsitemap — Method

findsitemap(unitcell, translation_operation)

Find which site gets mapped to which site with what lattice displacement. Return a vector of tuples (j, Rj) at index i, representing that the site i gets mapped to site j with lattice displacement Rj.

Since periodic boundary condition is always assumed, the return value always exists.

This function is used to find the mapping between sites under translation symmetry in a given lattice. The overall transformation will be a combiation of (1) the result here, and (2) Bravais lattice transformation. Since all the integer translations will be taken care of by the Bravais lattice transformation, Rj here is all zero vectors.

TightBindingLattice.findsitemap — Method

findsitemap(unitcell, translation_symmetry)

Find the sitemap for every element of the translation symmetry. Return [findsitemap(unitcell, m) for m in elements(translation_symmetry)].

TightBindingLattice.fract2carte — Method

fract2carte

Arguments

latticevectors ::AbstractArray{<:Real, 2}: square matrix whose columns are lattice vectors.

fc ::FractCoord: fractional coordinates

TightBindingLattice.fract2carte — Method

fract2carte

Arguments

latticevectors ::Array{Float64, 2}

fc ::FractCoord

TightBindingLattice.generate_subgroup — Method

generate_subgroup(group::FiniteGroup, idx::Integer)

subgroup generated by generators. ⟨ {g} ⟩

TightBindingLattice.generate_subgroup — Method

generate_subgroup(group::FiniteGroup, generators::Integer)

subgroup generated by generators. ⟨ S ⟩

TightBindingLattice.generator_elements — Method

generator_elements(sym::SymmorphicSymmetry)

TightBindingLattice.generator_indices — Method

generator_indices(sym::SymmorphicSymmetry)

TightBindingLattice.get_irrep_components — Method

get_irrep_components(sym::SymmorphicSymmetryEmbedding)

TightBindingLattice.get_irrep_components — Method

get_irrep_components(sym::SymmorphicSymmetry)

TightBindingLattice.get_irrep_components — Method

get_irrep_components(sym::SymmetryOrEmbedding)

Return a generator which gives IrrepComponent(sym, irrep_index, irrep_component).

TightBindingLattice.get_irrep_iterator — Method

get_irrep_iterator(sic::IrrepComponent)

Return a generator which gives (element, amplitude), where amplitude is the irrep component of element.

TightBindingLattice.get_irrep_iterator — Method

get_irrep_iterator(ssic::SymmorphicIrrepComponent)

TightBindingLattice.getsite — Method

getsite

Arguments

uc ::UnitCell{T}

index ::Integer

Return

(sitename, fractcoord)

TightBindingLattice.getsite — Method

getsite

Get the site (its site name and its fractional coordinates) with the given name.

Arguments

uc ::UnitCell{O}

name ::O

Return

(sitename, fractcoord)

TightBindingLattice.getsitecoord — Method

getsitecoord

Arguments

uc ::UnitCell

idx ::Integer

Return

fractcoord

TightBindingLattice.getsitecoord — Method

getsitecoord

Get the fractional coordinates of the site with the given name.

Arguments

uc ::UnitCell{O}

name ::O

Return

fractcoord

TightBindingLattice.getsiteindex — Method

getsiteindex

Get index of the given site.

Arguments

uc ::UnitCell{O}

name ::O

TightBindingLattice.getsiteindexcoord — Method

getsiteindexcoord

Arguments

uc ::UnitCell{T}

name ::T

Return

(index, fractcoord)

TightBindingLattice.getsitename — Method

getsitename

Arguments

uc ::UnitCell

index ::Integer

Return

sitename

TightBindingLattice.group_inverse — Method

group_inverse(group, g)

Get inverse of element/elements g.

TightBindingLattice.group_inverse — Method

group_inverse(group)

Get a function which gives inverse.

TightBindingLattice.group_isomorphism — Method

group_isomorphism(group1, group2)

Find the isomorphism ϕ: G₁ → G₂. Return nothing if not isomorphic.

TightBindingLattice.group_multiplication_table — Method

group_multiplication_table(group)

Return multiplcation table of the group.

TightBindingLattice.group_multiplication_table — Method

group_multiplication_table(elements, product=(*))

Generate a multiplication table from elements with product.

TightBindingLattice.group_order — Method

group_order(group, g)

Order of group element (i.e. period length)

TightBindingLattice.group_order — Method

group_order(group)

Order of group (i.e. number of elements)

TightBindingLattice.group_order — Method

group_order(sic::IrrepComponent)

Get order of the symmetry group of sic.

TightBindingLattice.group_product — Method

group_product(group, lhs, rhs)

Return the result of group multiplication of lhs and rhs. If both lhs and rhs are integers, return an integer. If either of them is a set (AbstractSet) of integers, then return a BitSet.

TightBindingLattice.group_product — Method

group_product(group)

Return a function which computes the group product.

TightBindingLattice.hassite — Method

hassite{T}

Test whether the unit cell contains the site of given name.

Arguments

uc ::UnitCell{O}

name ::O

TightBindingLattice.isabelian — Method

isabelian(group)

Check if the group is abelian.

TightBindingLattice.isbragg — Method

isbragg(k, t)

Check for Bragg condition at momentum k and translation t. Fractional momentum is normalized to 1, i.e. lies within [0, 1)ⁿ

TightBindingLattice.isbragg — Method

isbragg(tsym, tsym_irrep_index, translations)

TightBindingLattice.isbragg — Method

isbragg(tsym, tsym_irrep_index, translation)

TightBindingLattice.iscompatible — Method

iscompatible(tsic, psym)

TightBindingLattice.iscompatible — Method

iscompatible(lattice, point_operation)

TightBindingLattice.iscompatible — Method

iscompatible(lattice, point_symmetry)

TightBindingLattice.iscompatible — Method

iscompatible(lattice::Lattice, ssym::SymmorphicSymmetry)

TightBindingLattice.iscompatible — Method

iscompatible(lattice, translation_operation)

TightBindingLattice.iscompatible — Method

iscompatible(lattice, translation_symmetry)

Test whether lattice and the symmetry are compatible. For translation symmetry, this means that the hypercubic lattice for both are the same.

TightBindingLattice.iscompatible — Method

iscompatible(orthocube, operator::PointOperation{<:Integer})

TightBindingLattice.iscompatible — Method

iscompatible(orthocube, symmetry::PointSymmetry)

TightBindingLattice.iscompatible — Method

iscompatible(orthocube, operator::TranslationOperation{<:Integer})

TightBindingLattice.iscompatible — Method

iscompatible(orthocube, symmetry::TranslationSymmetry)

TightBindingLattice.iscompatible — Method

iscompatible(psym::PointSymmetry, tsym::TranslationSymmetry)

TightBindingLattice.iscompatible — Method

iscompatible(tsymbed, tsym_irrep_index, psymbed)

Check whether the point symmetry embedding psymbed is compatible with the translation symmetry irrep defined by tsym_irrep_index and symmetry(tsymbed). In other words, the little group elements

TightBindingLattice.iscompatible — Method

iscompatible(tsymbed, psymbed)

Check whether the point symmetry embedding psymbed is compatible with the translation symmetry embedding tsymbed, i.e. whether they have the same "lattice".

TightBindingLattice.iscompatible — Method

iscompatible(tsym, tsym_irrep_index, point_operation)

TightBindingLattice.iscompatible — Method

iscompatible(tsym, tsym_irrep_index, psym)

TightBindingLattice.iscompatible — Method

iscompatible(tsym::TranslationSymmetry, pop::PointOperation{<:Integer})

TightBindingLattice.iscompatible — Method

iscompatible(tsym::TranslationSymmetry, psym::PointSymmetry)

TightBindingLattice.ishomomorphic — Method

ishomomorphic(group, representation; product=(*), equal=(==))

Check whether representation is homomorphic to group under product and equal, order preserved.

TightBindingLattice.isidentity — Method

isidentity(t::TranslationOperation)

Return true if the translation operation is an identity, i.e. iszero(t.displacement)

TightBindingLattice.ispoint — Method

ispoint(arg::TranslationOperation)

Return true if arg is a point operation. The only way this can be true is when arg is an identity operation.

TightBindingLattice.issubgroup — Method

issubgroup(group, subset)

Check whether the given subset is a subgroup of group.

TightBindingLattice.istranslation — Method

istranslation(arg::TranslationOperation)

Always return true, since arg is already a translation operation.

TightBindingLattice.linpath — Method

momentumpath

Generate a list of momenta

Arguments

anchorpoints

(Optional) nseg - number of points in each segment

TightBindingLattice.little_group — Method

little_group(tsic, psym)

Return the FiniteGroup object that corresponds to the little group of psym at tsic.

TightBindingLattice.little_group — Method

little_group(tsym, tsym_irrep_index, psym)

TightBindingLattice.little_group — Method

little_group(tsym, psym, elements)

Generate a little group with given elements. The elements of the little group, which may be sparse, are compressed into consecutive integers.

TightBindingLattice.little_group_elements — Method

little_group_elements(tsic, psym)

Get the little group elements (i.e. indices) of psym corresponding to the irrep of translation symmetry specified by tsic. tsic and psym are either

IrrepComponent{TranslationSymmetry} and PointSymmetry, or

IrrepComponent{SymmetryEmbedding{TranslationSymmetry}} and SymmetryEmbedding{PointSymmetry}

TightBindingLattice.little_group_elements — Method

little_group_elements(tsymbed, tsym_irrep_index, psymbed)

TightBindingLattice.little_group_elements — Method

little_group_elements(tsymbed, psymbed)

TightBindingLattice.little_group_elements — Method

little_group_elements(tsym, tsym_irrep_index, psym)

Return elements of psym compatible with the translation symmetry tsym at irrep tsym_irrep_index

TightBindingLattice.little_group_elements — Method

little_group_elements(tsym, psym)

Return elements of psym compatible with the translation symmetry tsym.

TightBindingLattice.little_symmetry — Method

little_symmetry(tsic, psym)

Return the PointSymmetry object that corresponds to the little group of psym at tsic.

TightBindingLattice.little_symmetry — Method

little_symmetry(tsymbed, tsym_irrep_index, psymbed)

TightBindingLattice.little_symmetry — Method

little_symmetry(tsymbed, psymbed)

TightBindingLattice.little_symmetry — Method

little_symmetry(tsym, tsym_irrep_index, psym)

TightBindingLattice.little_symmetry — Method

little_symmetry(tsym, psym)

TightBindingLattice.little_symmetry_iso — Method

little_symmetry_iso(tsym, tsym_irrep_index, psym)

Find little symmetry using group isomorphism

TightBindingLattice.make_unitcell — Method

UnitCell

Construct an n-dimensional lattice.

Arguments

latticevectors ::AbstractArray{<:AbstractFloat, 2}: Lattice vectors

SiteType::DataType

Optional Arguments

tol=√ϵ: Epsilon

TightBindingLattice.make_unitcell — Method

UnitCell

Construct a one-dimensional lattice.

Arguments

latticeconstant::Real: Lattice constant

SiteType: List of sites

Optional Arguments

tol=√ϵ: Tolerance

TightBindingLattice.minimal_generating_set — Method

minimal_generating_set(group)

Get minimally generating set of the finite group.

TightBindingLattice.numsite — Method

numsite

Number of sites of the unit cell.

Arguments

uc ::UnitCell

TightBindingLattice.period_length — Method

period_length(group, g)

Order of group element (i.e. period length)

TightBindingLattice.project — Method

project(psym, projection; tol=√ϵ)

TightBindingLattice.read_point_symmetry — Method

read_point_symmetry(data::AbstractDict)

TightBindingLattice.sitecount — Method

sitecount

Number of sites of the unit cell.

Arguments

uc ::UnitCell

TightBindingLattice.symmetry_product — Method

symmetry_product(psym::PointSymmetry)

TightBindingLattice.symmetry_product — Method

symmetry_product(tsym::TranslationSymmetry)

TightBindingLattice.symmetry_product — Method

symmetry_product(sym::SymmorphicSymmetry{TranslationSymmetry,PointSymmetry,S}) where {S<:SpaceOperation{<:Integer}}

TightBindingLattice.translation_symmetry_embedding — Method

translation_symmetry_embedding(lattice::Lattice)

TightBindingLattice.whichunitcell — Method

whichunitcell

Return

Vector{Int}: the integer coordinate of the unitcell that the specified site/cartesian coordinate belongs to.

TightBindingLattice.whichunitcell — Method

whichunitcell

Return

Vector{Int}: the integer coordinate of the unitcell that the specified site/fractional coordinate belongs to.